fix: vanilla flag id not recognized
When the game calls the TakeFlag function, it uses the GetIdOffFlagInInventory function, which in vanilla returns the flag number. With Winch, the GetId... returns a string, with the full in game item id (so for example "flag-4")
The boolean value stored becomes "has-flag-flag4", which the game doesn't recognize using the GetHasFlag function.

This function replaces the start of the id with nothing to ensure vanilla flags are rightly recognized by the game